Network doesn't work.

+++ src/main/java/com/degoos/processing/game/network/ServerClient.java
@@ -47,7 +47,7 @@ public class ServerClient {
 			try {
 				while (!socket.isClosed()) {
 					if (inputStream.available() == 0) continue;
-					short s = inputStream.readShort();
+					short s = (short) inputStream.readInt();
 					Class<? extends Packet> clazz = Game.getPacketMap().get(s);
 					if (clazz == null) {
 						System.out.println("CLASS" + s + " NULL!");
@@ -96,7 +96,7 @@ public class ServerClient {
 	private void sendPacket(Packet packet, boolean secondTry) {
 		if (!secondTry) Engine.getEventManager().callEvent(new PacketSendEvent(packet, this));
 		try {
-			outputStream.writeShort(Game.getPacketMap().getPacketId(packet));
+			outputStream.writeInt(Game.getPacketMap().getPacketId(packet));
 			packet.write(outputStream);
 		} catch (Exception ex) {
 			ex.printStackTrace();
