+++ src/main/java/com/degoos/processing/game/network/ServerConnection.java
@@ -44,7 +44,7 @@ public class ServerConnection {
 					try {
 						while (!socket.isClosed()) {
 							if (inputStream.available() == 0) continue;
-							short s = inputStream.readShort();
+							short s = (short) inputStream.readInt();
 							Class<? extends Packet> clazz = Game.getPacketMap().get(s);
 							if (clazz == null) {
 								System.out.println("CLASS" + s + " NULL!");
@@ -84,7 +84,7 @@ public class ServerConnection {
 	private void sendPacket(Packet packet, boolean secondTry) {
 		if (!secondTry) Engine.getEventManager().callEvent(new ClientPacketSendEvent(packet));
 		try {
-			outputStream.writeShort(Game.getPacketMap().getPacketId(packet));
+			outputStream.writeInt(Game.getPacketMap().getPacketId(packet));
 			packet.write(outputStream);
 		} catch (Exception ex) {
 			ex.printStackTrace();
